refactor a bit, hook up open file for context listing and context updates
