+++ package.json
@@ -27,6 +27,7 @@
     "ollama": "^0.5.15",
     "open": "^10.2.0",
     "openai": "^5.3.0",
+    "pdfjs-dist": "^5.3.93",
     "tsx": "^4.19.2",
     "winston": "^3.17.0"
   },
